lib: Move the file length retrieval to platform code
The current code uses the std::io::File metadata method to retrieve the
underlying file length.

However, we're going to support RawFd on Unix systems, so we won't be
able to rely on the platform-agnostic File methods.

Let's move the function to the platform specific code, but still rely on
the old behaviour for now.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>