unix: Rely on fstat to get the backing file length
As part of our effort to move the unix implementation to rely on RawFd,
we'll need to move over from the metadata() method on std::fs::File to
something that could use a RawFd directly.

fstat() is provided by the libc crate we already depend on, so we can
just use it.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>