lib: Introduce an intermediate file representation
Since we want to be more flexible on the arguments to the mapping
function and allow to map a file descriptor without a backing file (such
as a buffer), we need to change slightly the function prototypes.

In order to preserve the API, an intermediate newtype is created that
only holds a reference to a file for now, and that newtype is passed to
all the pub functions that used to take a reference to a std::fs::File.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>