Support RawFd on cfg(unix)
Now that we have the infrastructure to introduce new source types for
the file descriptor or handle we want to map in memory, we can introduce
RawFd on Unix platforms.

Since it's not clear to me if Windows can allocate mappable file handles
without a backing file, the Windows implementation is kept as is for the
sake of simplicity.

Signed-off-by: Maxime Ripard <maxime@cerno.tech>